refactor: ui/write 패키지 분리
