+++ presentation/src/main/java/com/cheocharm/presentation/ui/write/diary/WriteFontAdapter.kt
@@ -1,4 +1,4 @@
-package com.cheocharm.presentation.ui.write
+package com.cheocharm.presentation.ui.write.diary
 
 import android.os.Bundle
 import android.view.View
