+++ presentation/src/main/java/com/cheocharm/presentation/ui/write/location/PictureFragment.kt
@@ -1,4 +1,4 @@
-package com.cheocharm.presentation.ui.write
+package com.cheocharm.presentation.ui.write.location
 
 import android.Manifest
 import android.app.Activity
