+++ presentation/src/main/res/layout/fragment_location.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewmodel"
-            type="com.cheocharm.presentation.ui.write.LocationViewModel" />
+            type="com.cheocharm.presentation.ui.write.location.LocationViewModel" />
 
     </data>
