+++ presentation/src/main/res/layout/write_align.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="viewmodel"
-            type="com.cheocharm.presentation.ui.write.WriteViewModel" />
+            type="com.cheocharm.presentation.ui.write.diary.WriteViewModel" />
 
     </data>
