+++ presentation/src/main/res/layout/write_color.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="viewmodel"
-            type="com.cheocharm.presentation.ui.write.WriteViewModel" />
+            type="com.cheocharm.presentation.ui.write.diary.WriteViewModel" />
 
     </data>
