+++ presentation/src/main/res/layout/write_font_family.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewmodel"
-            type="com.cheocharm.presentation.ui.write.WriteFontViewModel" />
+            type="com.cheocharm.presentation.ui.write.diary.WriteFontViewModel" />
 
     </data>
