+++ presentation/src/main/res/layout/write_font_size.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="viewmodel"
-            type="com.cheocharm.presentation.ui.write.WriteFontViewModel" />
+            type="com.cheocharm.presentation.ui.write.diary.WriteFontViewModel" />
 
     </data>
